# Optimize remove_audio, resize, rotate