--- a/augly/video/augmenters/ffmpeg/no_audio.py
+++ b/augly/video/augmenters/ffmpeg/no_audio.py
@@ -1,22 +1,35 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-from typing import Tuple, Dict
+from typing import List, Optional
 
-from augly.video.augmenters.ffmpeg.base_augmenter import BaseFFMPEGAugmenter
-from ffmpeg.nodes import FilterableStream
+from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
 
 
-class VideoAugmenterByRemovingAudio(BaseFFMPEGAugmenter):
-    def add_augmenter(
-        self, in_stream: FilterableStream, **kwargs
-    ) -> Tuple[FilterableStream, Dict]:
+class VideoAugmenterByRemovingAudio(BaseVidgearFFMPEGAugmenter):
+    def get_command(
+        self, video_path: str, output_path: Optional[str] = None, **kwargs
+    ) -> List[str]:
         """
         Removes the audio from the video
 
-        @param in_stream: the FFMPEG object of the video
+        @param video_path: the path to the video to be augmented
 
-        @returns: a tuple containing the FFMPEG object with the augmentation
-            applied and a dictionary with any output arguments as necessary
+        @param output_path: the path in which the resulting video will be stored.
+            If not passed in, the original video file will be overwritten
+
+        @returns: a list of strings of the FFMPEG command if it were to be written
+            in a command line
         """
-        return in_stream.video, {"an": None}
+        command = [
+            "-y",
+            "-i",
+            video_path,
+            "-c",
+            "copy",
+            "-an",
+            "-preset",
+            "ultrafast",
+            output_path,
+        ]
+        return command