--- a/augly/video/augmenters/ffmpeg/no_audio.py
+++ b/augly/video/augmenters/ffmpeg/no_audio.py
@@ -1,15 +1,13 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-from typing import List, Optional
+from typing import List
 
 from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
 
 
 class VideoAugmenterByRemovingAudio(BaseVidgearFFMPEGAugmenter):
-    def get_command(
-        self, video_path: str, output_path: Optional[str] = None, **kwargs
-    ) -> List[str]:
+    def get_command(self, video_path: str, output_path: str) -> List[str]:
         """
         Removes the audio from the video
 
@@ -32,4 +30,5 @@ class VideoAugmenterByRemovingAudio(BaseVidgearFFMPEGAugmenter):
             "ultrafast",
             output_path,
         ]
+
         return command