--- a/augly/video/augmenters/ffmpeg/rotate.py
+++ b/augly/video/augmenters/ffmpeg/rotate.py
@@ -2,26 +2,40 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 
 import math
-from typing import Tuple, Dict
+from typing import List, Optional
 
-from augly.video.augmenters.ffmpeg.base_augmenter import BaseFFMPEGAugmenter
-from ffmpeg.nodes import FilterableStream
+from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
 
 
-class VideoAugmenterByRotation(BaseFFMPEGAugmenter):
+class VideoAugmenterByRotation(BaseVidgearFFMPEGAugmenter):
     def __init__(self, degrees: float):
         assert isinstance(degrees, (float, int)), "Expected 'degrees' to be a number"
         self.degrees = degrees
 
-    def add_augmenter(
-        self, in_stream: FilterableStream, **kwargs
-    ) -> Tuple[FilterableStream, Dict]:
+    def get_command(
+        self, video_path: str, output_path: Optional[str] = None, **kwargs
+    ) -> List[str]:
         """
         Rotates the video
 
-        @param in_stream: the FFMPEG object of the video
+        @param video_path: the path to the video to be augmented
 
-        @returns: a tuple containing the FFMPEG object with the augmentation
-            applied and a dictionary with any output arguments as necessary
+        @param output_path: the path in which the resulting video will be stored.
+            If not passed in, the original video file will be overwritten
+
+        @returns: a list of strings of the FFMPEG command if it were to be written
+            in a command line
         """
-        return in_stream.video.filter("rotate", self.degrees * (math.pi / 180)), {}
+        command = [
+            "-y",
+            "-i",
+            video_path,
+            "-vf",
+            f"rotate={self.degrees * (math.pi / 180)}",
+            "-c:a",
+            "copy",
+            "-preset",
+            "ultrafast",
+            output_path,
+        ]
+        return command