## Summary
- [X] I have read CONTRIBUTING.md to understand how to contribute to this repository :)

<Please summarize what you are trying to achieve, what changes you made, and how they acheive the desired result.>
I'm trying to speed up the following video augmentations: remove_audio, resize, and rotate.

Here are the times which I measured with time for i in (seq 5); python -m unittest augly.tests.video_tests.transforms.ffmpeg_test.TransformsVideoUnitTest.test_{respective_augmentation};; end

Runtime with changes for RemoveAudio: 3.69s (test finished about 0.4s)
Runtime without changes for RemoveAudio: 16.468s (test finished about 12s)

Runtime with changes for Resize: 3.76s (test finished about 1.5s)
Runtime without changes for Resize: 15.52s (test finished about 11s)

Runtime with changes for Rotate: 5s (test finished about 5.5s)
Runtime without changes for Rotate: 15.024s (test finished about 12s)

Since I changed resize, pixelization and random_pixelization got affected so here are the times as well

Runtime with changes for Pixelization: 5.216s (test finished about 2s)
Runtime without changes for Pixelization: 10.322s (test finished about 7s)

Runtime with changes for RandomPixelization: 5.286s (test finished about 2.5s)
Runtime without changes for RandomPixelization: 10.56s (test finished about 6.5s)

I've also added an additional check in `augly/video/augmenter/ffmpeg/base_augmenter.py` under the VidGear base class add_augmenter function for if output_path and video_path are the same. Since we are executing a ffmpeg command we can't do this because we can't write over the file that we are reading from since ffmpeg doesn't put a file in some buffer, we have to create a temporary file.



## Unit Tests
If your changes touch the `audio` module, please run all of the `audio` tests and paste the output here. Likewise for `image`, `text`, & `video`. If your changes could affect behavior in multiple modules, please run the tests for all potentially affected modules. If you are unsure of which modules might be affected by your changes, please just run all the unit tests.

### Audio
```bash
python -m unittest discover -s augly/tests/audio_tests/ -p "*"
```

### Image
```bash
python -m unittest discover -s augly/tests/image_tests/ -p "*_test.py"
# Or `python -m unittest discover -s augly/tests/image_tests/ -p "*.py"` to run pytorch test too (must install `torchvision` to run)
```

### Text
```bash
python -m unittest discover -s augly/tests/text_tests/ -p "*"
```

### Video
```bash
python -m unittest discover -s augly/tests/video_tests/ -p "*"
```

### All
```bash
python -m unittest discover -s augly/tests/ -p "*"
```

## Other testing

If applicable, test your changes and paste the output here. For example, if your changes affect the requirements/installation, then test installing augly in a fresh conda env, then make sure you are able to import augly & run the unit test